complete public reducer
